pyproject.toml
[tool.poetry]
name = "skypro-diploma"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
black = "^24.8.0"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
setuptools = "^74.1.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-filter = "^24.3"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
